add named embeddings and fix inference

+++ pykg2vec/utils/generator.py
@@ -248,12 +248,13 @@ def process_function_multiclass(raw_queue, processed_queue, config):
         # It looks Torch sparse tensor does not work in multi processing
         # so they need to be converted to dense, which is not memory efficient
         # https://github.com/pytorch/pytorch/pull/27062
-        hr_t = torch.sparse.FloatTensor(torch.LongTensor(_to_sparse_i(indices_hr_t)), values_hr_t, torch.Size(shape)).to_dense()
-        tr_h = torch.sparse.FloatTensor(torch.LongTensor(_to_sparse_i(indices_tr_h)), values_tr_h, torch.Size(shape)).to_dense()
+        # https://github.com/pytorch/pytorch/issues/20248
+        hr_t = torch.sparse.LongTensor(torch.LongTensor(_to_sparse_i(indices_hr_t)), values_hr_t, torch.Size(shape)).to_dense()
+        tr_h = torch.sparse.LongTensor(torch.LongTensor(_to_sparse_i(indices_tr_h)), values_tr_h, torch.Size(shape)).to_dense()
 
         if neg_rate > 0:
-            neg_hr_t = torch.sparse.FloatTensor(torch.LongTensor(_to_sparse_i(neg_indices_hr_t)), neg_values_hr_t, torch.Size(shape)).to_dense()
-            neg_tr_h = torch.sparse.FloatTensor(torch.LongTensor(_to_sparse_i(neg_indices_tr_h)), neg_values_tr_h, torch.Size(shape)).to_dense()
+            neg_hr_t = torch.sparse.LongTensor(torch.LongTensor(_to_sparse_i(neg_indices_hr_t)), neg_values_hr_t, torch.Size(shape)).to_dense()
+            neg_tr_h = torch.sparse.LongTensor(torch.LongTensor(_to_sparse_i(neg_indices_tr_h)), neg_values_tr_h, torch.Size(shape)).to_dense()
         
             hr_t = hr_t.add(neg_hr_t)
             tr_h = tr_h.add(neg_tr_h)
